JS to make video (on index page) play on hover
Poster image shows when page loads, when hovered over, first frame of video shows. When clicked play on the controls, video plays but only if you keep the mouse over the element.  Move cursor away and it pauses, hover back on and it plays. Throws Javascript error "Uncaught (in promise) DOMException"